added footstep files and fixed audio bug
ma_sound* would get sent to the engine, but then it would get destroyed when the SoundArray tried to add a new sound to its std::vector and that deleted the pointer the Audio object held, leaving the engine with an invalid pointer. To fix I added a move constructor to the Audio class.